+++ src/components/About/Banner.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import "./About.css";
+
+//Banner
+const Banner = () => {
+  return (
+    <header class="header">
+    <div class="header__text-box">
+      <h1 class="heading-primary">
+        <span class="heading-primary__main">ABOUT US</span>
+      </h1>
+
+      <a href="#" class="btn_ btn--white btn--animated">
+        Discover Who We Are?
+      </a>
+    </div>
+  </header>
+  )
+}
+
+export default Banner
